Fixing another problem caused by restricting the exceptions caught in remove_file.
git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@7895 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 9f92fb5df5bef4cdf8adaa28e6572fdf58744212

+++ tv/portable/iconcache.py
@@ -229,7 +229,8 @@ class IconCache:
                 self.remove_file(tmp_filename)
                 return
 
-            self.remove_file(self.filename)
+            if self.filename:
+                self.remove_file(self.filename)
 
             # Create a new filename always to avoid browser caching in case a 
             # file changes.
